[android] Separated content frame aka menu from button lines frame
